RTK_project/requirements.txt
requests==2.31.0
pytest==7.4.0
pytest-check==2.2.5
pytest-xdist==3.3.1
allure-pytest==2.8.18